# Conflicts:
#	cpp/subprojects/boosting/include/mlrl/boosting/util/blas.hpp
#	cpp/subprojects/boosting/include/mlrl/boosting/util/lapack.hpp
#	cpp/subprojects/common/include/mlrl/common/sampling/random.hpp

+++ cpp/subprojects/seco/src/mlrl/seco/heuristics/heuristic_wra.cpp
@@ -10,8 +10,8 @@ namespace seco {
     class Wra final : public IHeuristic {
         public:
 
-            float64 evaluateConfusionMatrix(float64 cin, float64 cip, float64 crn, float64 crp, float64 uin,
-                                            float64 uip, float64 urn, float64 urp) const override {
+            float32 evaluateConfusionMatrix(float32 cin, float32 cip, float32 crn, float32 crp, float32 uin,
+                                            float32 uip, float32 urn, float32 urp) const override {
                 return wra(cin, cip, crn, crp, uin, uip, urn, urp);
             }
     };
