+++ cpp/subprojects/seco/src/mlrl/seco/lift_functions/lift_function_kln.cpp
@@ -4,8 +4,8 @@
 
 namespace seco {
 
-    static inline float64 calculateLiftInternally(uint32 numLabels, float64 k) {
-        return 1.0 + (k * std::log((float64) numLabels));
+    static inline float32 calculateLiftInternally(uint32 numLabels, float32 k) {
+        return 1.0 + (k * std::log((float32) numLabels));
     }
 
     /**
@@ -15,9 +15,9 @@ namespace seco {
     class KlnLiftFunction final : public ILiftFunction {
         private:
 
-            const float64 k_;
+            const float32 k_;
 
-            const float64 maxLift_;
+            const float32 maxLift_;
 
         public:
 
@@ -26,13 +26,13 @@ namespace seco {
              *                  values for "k". Must be at least 0
              * @param maxLift   The maximum possible lift
              */
-            KlnLiftFunction(float64 k, float64 maxLift) : k_(k), maxLift_(maxLift) {}
+            KlnLiftFunction(float32 k, float32 maxLift) : k_(k), maxLift_(maxLift) {}
 
-            float64 calculateLift(uint32 numLabels) const override {
+            float32 calculateLift(uint32 numLabels) const override {
                 return calculateLiftInternally(numLabels, k_);
             }
 
-            float64 getMaxLift(uint32 numLabels) const override {
+            float32 getMaxLift(uint32 numLabels) const override {
                 return maxLift_;
             }
     };
@@ -44,9 +44,9 @@ namespace seco {
     class KlnLiftFunctionFactory final : public ILiftFunctionFactory {
         private:
 
-            const float64 k_;
+            const float32 k_;
 
-            const float64 maxLift_;
+            const float32 maxLift_;
 
         public:
 
@@ -55,7 +55,7 @@ namespace seco {
              * @param k         The value of the parameter "k". The steepness of the lift function increases with larger
              *                  values for "k". Must be at least 0
              */
-            KlnLiftFunctionFactory(uint32 numLabels, float64 k)
+            KlnLiftFunctionFactory(uint32 numLabels, float32 k)
                 : k_(k), maxLift_(calculateLiftInternally(numLabels, k)) {}
 
             std::unique_ptr<ILiftFunction> create() const override {
@@ -63,14 +63,14 @@ namespace seco {
             }
     };
 
-    KlnLiftFunctionConfig::KlnLiftFunctionConfig() : k_(0.2) {}
+    KlnLiftFunctionConfig::KlnLiftFunctionConfig() : k_(0.2f) {}
 
-    float64 KlnLiftFunctionConfig::getK() const {
+    float32 KlnLiftFunctionConfig::getK() const {
         return k_;
     }
 
-    IKlnLiftFunctionConfig& KlnLiftFunctionConfig::setK(float64 k) {
-        util::assertGreater<float64>("k", k, 0);
+    IKlnLiftFunctionConfig& KlnLiftFunctionConfig::setK(float32 k) {
+        util::assertGreater<float32>("k", k, 0);
         k_ = k;
         return *this;
     }
